replace false_type/true_type with named tags

+++ include/dctl/ui/dxp/message.hpp
@@ -24,8 +24,8 @@ class Message
         bool invariant() const
         {
                 return
-                        static_cast<int>(do_header().length()) == header_length_ &&
-                        static_cast<int>(do_body().length()) <= max_body_length_
+                        do_header().length() == header_length_ &&
+                        do_body().length() <= max_body_length_
                 ;
         }
